Add a test case for the useEffect in useSizeHandlers
