+++ src/components/forms/model-form/model-form.js
@@ -70,11 +70,6 @@ const ModelForm = ({ model, id, isEdit }) => {
     categories: Categories.categories
   }));
 
-  const { sizes, onSizeSubmit, onSizeDelete } = useSizeHandlers(model.sizes);
-
-  const [sizesTouched, setSizesTouched] = useState(false);
-  const sizesAdded = sizes.map((size) => size.name);
-
   const [category, setCategory] = useState(model.category._id || '');
   const {
     values,
@@ -115,18 +110,18 @@ const ModelForm = ({ model, id, isEdit }) => {
     }
   });
 
+  const { sizes, onSizeSubmit, onSizeDelete } = useSizeHandlers(
+    model.sizes,
+    setFieldValue
+  );
+  const [sizesTouched, setSizesTouched] = useState(false);
+  const sizesAdded = sizes.map((size) => size.name);
+
   const sizeUtils = {
     onSizeSubmit,
     onSizeDelete,
     sizesAdded
   };
-
-  useEffect(() => {
-    const updatedSizes = sizes.map(({ _id, ...size }) =>
-      _id.includes('size_') ? size : { _id, ...size }
-    );
-    setFieldValue('sizes', updatedSizes);
-  }, [sizes, setFieldValue]);
 
   const [sizeFormExpanded, setSizeFormExpanded] = useState('');
