+++ src/hooks/model-form/use-size-handlers.js
@@ -1,6 +1,6 @@
-import { useState } from 'react';
+import { useState, useEffect } from 'react';
 
-const useSizeHandlers = (initialSizes) => {
+const useSizeHandlers = (initialSizes, setFieldValue) => {
   const [sizes, setSizes] = useState(initialSizes || []);
 
   const onSizeSubmit = (newSize) => {
@@ -20,6 +20,13 @@ const useSizeHandlers = (initialSizes) => {
       currentSizes.filter((size) => size._id !== sizeIdToDelete)
     );
   };
+
+  useEffect(() => {
+    const updatedSizes = sizes.map(({ _id, ...size }) =>
+      _id.includes('size_') ? size : { _id, ...size }
+    );
+    setFieldValue('sizes', updatedSizes);
+  }, [sizes, setFieldValue]);
 
   return {
     sizes,
