+++ src/hooks/model-form/use-size-handlers.spec.js
@@ -1,3 +1,4 @@
+import React from 'react';
 import { renderHook, act } from '@testing-library/react-hooks';
 import useSizeHandlers from './use-size-handlers';
 import {
@@ -8,8 +9,14 @@ import {
 } from './use-size-handlers.variables';
 
 describe('useSizeHandlers hook test', () => {
+  const mockUseEffect = jest.spyOn(React, 'useEffect');
+
+  beforeEach(() => {
+    mockUseEffect.mockImplementation(() => jest.fn());
+  });
+
   it('Should add a size to the sizes state variable', () => {
-    const { result } = renderHook(() => useSizeHandlers(initialSizes));
+    const { result } = renderHook(() => useSizeHandlers(initialSizes, null));
 
     act(() => {
       result.current.onSizeSubmit(newSize);
@@ -18,8 +25,18 @@ describe('useSizeHandlers hook test', () => {
     expect(result.current.sizes[1]).toEqual(newSize);
   });
 
+  it('Should call useEffect after the size addition', () => {
+    const { result } = renderHook(() => useSizeHandlers(initialSizes, null));
+
+    act(() => {
+      result.current.onSizeSubmit(newSize);
+    });
+
+    expect(mockUseEffect).toHaveBeenCalledTimes(2);
+  });
+
   it('Should update a size in the sizes state variable', () => {
-    const { result } = renderHook(() => useSizeHandlers(initialSizes));
+    const { result } = renderHook(() => useSizeHandlers(initialSizes, null));
 
     act(() => {
       result.current.onSizeSubmit({
@@ -32,7 +49,7 @@ describe('useSizeHandlers hook test', () => {
   });
 
   it('Should delete a size from the sizes state variable', () => {
-    const { result } = renderHook(() => useSizeHandlers(initialSizes));
+    const { result } = renderHook(() => useSizeHandlers(initialSizes, null));
 
     act(() => {
       result.current.onSizeDelete(sizeToUpdate._id);
